+++ concordium-consensus/src/Concordium/Scheduler/Environment.hs
@@ -356,7 +356,7 @@ class
     -- * Chain updates
 
     -- | Get the current authorized keys for updates.
-    getUpdateKeyCollection :: m (UpdateKeysCollection (AuthorizationsVersionForPV (MPV m)))
+    getUpdateKeyCollection :: m (UpdateKeysCollection (AuthorizationsVersionFor (MPV m)))
 
     -- | Get the next sequence number of updates of a given type.
     getNextUpdateSequenceNumber :: UpdateType -> m UpdateSequenceNumber
@@ -366,7 +366,7 @@ class
     --  The next sequence number will be correspondingly incremented,
     --  and any queued updates of the given type with a later effective
     --  time are cancelled.
-    enqueueUpdate :: TransactionTime -> UpdateValue (ChainParametersVersionFor (MPV m)) -> m ()
+    enqueueUpdate :: TransactionTime -> UpdateValue (ChainParametersVersionFor (MPV m)) (AuthorizationsVersionFor (MPV m)) -> m ()
 
     -- | Increment the update sequence number for Protocol Level Tokens (PLT).
     -- Unlike the other chain updates this is a separate function,